pyproject.toml
[tool.poetry]
name = "20-lesson-hw"
version = "0.1.0"
description = ""
authors = ["NastyaGavrichenkova <lady_markos@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.2"
requests = "^2.31.0"
allure-pytest = "^2.13.2"
curlify = "^2.2.1"
jsonschema = "^4.19.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
